hdf4 bugs

+++ cdm/src/main/java/ucar/nc2/iosp/LayoutBB.java
@@ -112,6 +112,6 @@ public interface LayoutBB {
      *
      * @return starting element in the array (Note: elements, not bytes)
      */
-    public int getDestElem();
+    public long getDestElem();
   }
 }
